Enhancement: Heavyweight Updates! Splicing javac commands without unpacking
