+++ pom.xml
@@ -13,7 +13,7 @@
     <groupId>io.github.bigwolftime</groupId>
     <artifactId>redefine</artifactId>
     <packaging>jar</packaging>
-    <version>0.0.6</version>
+    <version>0.0.7</version>
 
     <name>redefine</name>
     <description>Use ByteBuddy API for code redefinition(Java)</description>
@@ -105,12 +105,6 @@
             <groupId>org.apache.commons</groupId>
             <artifactId>commons-io</artifactId>
             <version>1.3.2</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-compress</artifactId>
-            <version>1.25.0</version>
         </dependency>
 
     </dependencies>
