+++ redefine-git.properties
@@ -1,13 +1,13 @@
 #Generated by Git-Commit-Id-Plugin
-git.build.version=0.0.5-RELEASE
-git.closest.tag.commit.count=
-git.closest.tag.name=
-git.commit.id=853dd21420945c7624b4983d156af287bc37d836
-git.commit.id.abbrev=853dd21
-git.commit.id.describe=853dd21
-git.commit.id.describe-short=853dd21
-git.commit.message.full=m
-git.commit.message.short=m
-git.commit.user.email=liuxin441@gmail.com
-git.commit.user.name=liuxin
-git.dirty=false
+git.build.version=0.0.7
+git.closest.tag.commit.count=0
+git.closest.tag.name=v0.0.6
+git.commit.id=1033945ad8c79e96fa89a4ae8857cdaf36117c34
+git.commit.id.abbrev=1033945
+git.commit.id.describe=v0.0.6-dirty
+git.commit.id.describe-short=v0.0.6-dirty
+git.commit.message.full=Merge pull request \#1 from bigwolftime/base64\n\nEnhancement\: support base64
+git.commit.message.short=Merge pull request \#1 from bigwolftime/base64
+git.commit.user.email=15247656141@163.com
+git.commit.user.name=bigwolftime
+git.dirty=true
