--- src/main/java/com/bigwolftime/constant/JarFileConstant.java
@@ -1,26 +1,0 @@
-package com.bigwolftime.constant;
-
-/**
- * @author liuxin
- */
-public class JarFileConstant {
-    private JarFileConstant() {}
-
-    /**
-     * root path
-     */
-    public static final String BOOT_INF = "BOOT-INF/";
-
-    /**
-     * classes directory
-     */
-    public static final String CLASSES = "classes";
-
-    /**
-     * dependencies directory
-     */
-    public static final String LIB = "lib";
-
-    public static final String UNCOMPRESS_PATH = "/tmp/redefine/uncompress/";
-
-}
